+++ src/Lambda/Notebook/Persistance/Header.hs
@@ -3,7 +3,7 @@ module Lambda.Notebook.Persistance.Header where
 import qualified Data.Map as M
 import qualified Data.UUID as U
 import Lambda.Notebook.Persistance.Model (Notebook)
-import Lambda.Notebook.Storage (InsertM)
+import Lambda.Notebook.Storage (ReadM, InsertM)
 
 -- api ------------------------------------------------------------------------
 
@@ -18,6 +18,9 @@ type LookupNotebookM m = U.UUID -> m (Maybe Notebook)
 
 -- | persist notebook under certain uuid
 type SaveNotebookM m = InsertM U.UUID Notebook m
+
+-- | list of available notebooks
+type ListNotebooksM m = ReadM [(U.UUID, Notebook)] m
 
 -- type classes ---------------------------------------------------------------